requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
beautifulsoup4>=4.12.0
requests>=2.32.0
pypdf>=4.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
langchain-core>=0.3.0
google-generativeai>=0.8.0
